requirements-edge.txt
opencv-python
numpy
scipy
pandas
pyyaml
onnxruntime
bleak
